feat: claim modal ui
